requirements.txt
streamlit
snowflake-connector-python
pandas
pydantic
langchain-google-genai
apache-airflow
requests
google-auth